move `maintest` package
